yield evolve with D(but did not work out)
The dissipated energy is getting smaller when the material reaches the
end of fatigue, because ||S-b|| projects on the yield limit in local
plastic zone, and ||S-b|| is the multiplier of dW.
In chaboche model, (1-D) is added to the denominator of equivalent
stress, the stress intensity gets bigger with increasing number of
cycles. Our model fits Chaboche law so the power of Smax and (1-D) is
the same in Wcyc. In addition, it is taken into account when we put dD
on one side.